Ajouter js standard Window confirm() Method avant supprimer une preuve

+++ src/app/proof/proof.component.ts
@@ -65,6 +65,9 @@ export class ProofComponent implements OnInit {
   }
 
   public onClickDelete(proof: any) {
-    this.proofService.deleteProof(proof);
+    const isConfirmed = confirm('Voulez-vous supprimer cette preuve "' + proof.name + '" ?');
+    if (isConfirmed) {
+      this.proofService.deleteProof(proof);
+    }
   }
 }
